fix(website): extract graph-utils lib to fix SSR/browser bundle leak
graph.ts (an SSR-only API route) was directly imported at runtime by
ArchitekturGraph.svelte (client:load), pulling node:fs/promises and
node:path into the browser bundle → fatal Rollup build error.

Fix 1: Move all browser-safe interfaces (GraphNode, GraphEdge, etc.) and
pure functions (buildStatusMap, resolveGraph, matchPodsToNode) to
src/lib/graph-utils.ts. graph.ts now only holds the GET handler and
re-exports from graph-utils. The 4 Svelte components now import from
graph-utils directly.

Fix 2: Change `import { resolve4 } from 'dns'` to default import
`import dns from 'dns'` in knowledge-db.ts, codesearch-db.ts, and
website-db.ts. Vite's browser stub for 'dns' is `{}` (no named exports),
so a named import fails at bundle time with "not exported"; a default
import is safe (returns `{}`) and avoids the fatal error.

Fixes the broken build-website.yml workflow (failing since PR #1617).
